# docs/swagger.yaml
basePath: /api/
definitions:
  auth.ChangePassword:
    properties:
      new_password:
        type: string
      old_password:
        type: string
    required:
    - new_password
    - old_password
    type: object
  auth.IdResponse:
    properties:
      id:
        type: string
    type: object
  auth.MessageResponse:
    properties:
      message:
        type: string
    type: object
  auth.SignInInput:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  auth.TokenResponse:
    properties:
      token:
        type: string
    type: object
  auth.UsernameInput:
    properties:
      username:
        type: string
    type: object
...
  responses.ErrorResponse:
    properties:
      message:
        type: string
    type: object
...
host: localhost:8000
info:
  contact:
    name: Tairyoku
    url: https://github.com/Tairyoku
  description: Серверна частина онлайн чату.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Server API
  version: 1.0.0
paths:
  /auth/change/icon:
    put:
      description: Користувач надсилає новий файл зображення. Замінює зображення на
        нове.
      produces:
      - application/json
      responses:
        "200":
          description: icon changed
          schema:
            $ref: '#/definitions/auth.MessageResponse'
        "404":
          description: incorrect user data
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: delete icon error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Change username
      tags:
      - auth
  /auth/change/password:
    put:
      consumes:
      - application/json
      description: |-
        Користувач надсилає поточний та новий паролі.
        Після перевірки правильності ведення поточного паролю, змінює пароль на новий.
      parameters:
      - description: actual and new password
        in: body
        name: passwords
        required: true
        schema:
          $ref: '#/definitions/auth.ChangePassword'
      produces:
      - application/json
      responses:
        "200":
          description: password changed
          schema:
            $ref: '#/definitions/auth.MessageResponse'
        "400":
          description: incorrect password
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: incorrect user data
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: update password error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Change user password
      tags:
      - auth
  /auth/change/username:
    put:
      consumes:
      - application/json
      description: |-
        Користувач надсилає новий нікнейм.
        Після перевірки нового нікнейму на унікальність, змінює нікнейм на новий.
      parameters:
      - description: New username
        in: body
        name: username
        required: true
        schema:
          $ref: '#/definitions/auth.UsernameInput'
      produces:
      - application/json
      responses:
        "200":
          description: username changed
          schema:
            $ref: '#/definitions/auth.MessageResponse'
        "400":
          description: incorrect request data
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: incorrect user data
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: update username error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Change username
      tags:
      - auth
  /auth/get-me:
    get:
      description: Отримує у header запиту токен, повертає ID користувача.
      produces:
      - application/json
      responses:
        "200":
          description: result is user ID
          schema:
            $ref: '#/definitions/auth.TokenResponse'
        "404":
          description: user not found
          schema:
            $ref: '#/definitions/auth.IdResponse'
      security:
      - ApiKeyAuth: []
      summary: Decoded user ID
      tags:
      - auth
  /auth/sign-in:
    post:
      consumes:
      - application/json
      description: |-
        Користувач відправляє ім'я та пароль.
        Сервер поверне token існуючого користувача або помилку якщо користувача не існує.
      parameters:
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/auth.SignInInput'
      produces:
      - application/json
      responses:
        "200":
          description: result is user token
          schema:
            $ref: '#/definitions/auth.TokenResponse'
        "400":
          description: incorrect request data
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: user not found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "409":
          description: incorrect password
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: check user error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      summary: Generate a new user token
      tags:
      - auth
  /auth/sign-up:
    post:
      consumes:
      - application/json
      description: |-
        Користувач відправляє ім'я та пароль.
        За отриманими даними буде створено нового користувача.
        Сервер поверне token нового користувача.
      operationId: add-new-user
      parameters:
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/auth.SignInInput'
      produces:
      - application/json
      responses:
        "200":
          description: result is user token
          schema:
            $ref: '#/definitions/auth.TokenResponse'
        "400":
          description: Password must be at least 6 symbols
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "409":
          description: username is already used
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: generate token error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      summary: Create a new user
      tags:
      - auth
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"